IT WORKS! Now working on making order not matter
It turns out that the problem wasn't with addChild at all. Rather, when
passing Parser's dictionary from parseLine() when creating new Elements,
not including "children" in that dictionary caused all Elements to just
assign their children lists to the same memory addresses. I'm sure I
won't understand this the next time I look at it and I'm sure I'm not
using the "description" field of this commit thing right but it works